requirements.txt
tweepy==4.14.0
pandas==2.1.0
transformers==4.34.0
spacy==3.7.0
arcgis
fastapi==0.109.1
uvicorn==0.23.2
python-dotenv==1.0.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0